# Reduce memory allocation for list mergers