--- a/pkg/merger/list_merger.go
+++ b/pkg/merger/list_merger.go
@@ -22,7 +22,21 @@ func (p Rid) Less(i, j int) bool { return p[i].Len() < p[j].Len() }
 func (p Rid) Swap(i, j int) { p[i], p[j] = p[j], p[i] }
 
 // MergeCandidate is result of merging Rid
-type MergeCandidate struct {
-	Position uint32
-	Overlap  int
+type MergeCandidate uint64
+
+func NewMergeCandidate(position uint32, overlap int) MergeCandidate {
+	return MergeCandidate(uint64(position) << 32 | uint64(overlap))
 }
+
+func (m MergeCandidate) Position() uint32 {
+	return uint32(m >> 32)
+}
+
+func (m MergeCandidate) Overlap() int {
+	return int(uint32(m))
+}
+
+func (m *MergeCandidate) Increment() {
+	*m = NewMergeCandidate(m.Position(), m.Overlap() + 1)
+}
+