--- a/pkg/merger/scan_count.go
+++ b/pkg/merger/scan_count.go
@@ -33,8 +33,8 @@ func (lm *scanCount) Merge(rid Rid, threshold int) ([]MergeCandidate, error) {
 		j, endMergeCandidate = 0, len(candidates)
 
 		for j < endMergeCandidate || isValid {
-			if j >= endMergeCandidate || (isValid && candidates[j].Position > current) {
-				tmp = append(tmp, MergeCandidate{current, 1})
+			if j >= endMergeCandidate || (isValid && candidates[j].Position() > current) {
+				tmp = append(tmp, NewMergeCandidate(current, 1))
 
 				if list.HasNext() {
 					current, err = list.Next()
@@ -45,11 +45,11 @@ func (lm *scanCount) Merge(rid Rid, threshold int) ([]MergeCandidate, error) {
 				} else {
 					isValid = false
 				}
-			} else if !isValid || (j < endMergeCandidate && candidates[j].Position < current) {
+			} else if !isValid || (j < endMergeCandidate && candidates[j].Position() < current) {
 				tmp = append(tmp, candidates[j])
 				j++
 			} else {
-				candidates[j].Overlap++
+				candidates[j].Increment()
 				tmp = append(tmp, candidates[j])
 				j++
 
@@ -71,7 +71,7 @@ func (lm *scanCount) Merge(rid Rid, threshold int) ([]MergeCandidate, error) {
 	tmp = tmp[:0]
 
 	for _, c := range candidates {
-		if c.Overlap >= threshold {
+		if c.Overlap() >= threshold {
 			tmp = append(tmp, c)
 		}
 	}