This reduces the memory allocation for `ListMerger` algorithms. We declare `MergeCandidate` as `uint64` type and store the `position`  in the first 32 bits and the `overlap` in the least. Here is some benchmarks:

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkWordsOnDisc-8     358533        317211        -11.53%

benchmark                  old allocs     new allocs     delta
BenchmarkWordsOnDisc-8     614            613            -0.16%

benchmark                  old bytes     new bytes     delta
BenchmarkWordsOnDisc-8     112773        65468         -41.95%

```